requirements.txt
streamlit>=1.32.0
openai>=1.30.0
mcp[cli]>=1.3.0
fastapi>=0.110.0
uvicorn>=0.29.0
trafilatura>=1.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
